Move the heavy client code to core.
