+++ example/tools/git/src/main/scala/GitUtil.scala
@@ -1,15 +1,8 @@
 import java.io.File
-import java.io.{InputStream, OutputStream}
 import java.nio.file.{Paths, Files, StandardCopyOption}
 import com.typesafe.config._
-import scala.util.{Try, Success, Failure}
-import scala.annotation.tailrec
-import scala.concurrent._
-import ExecutionContext.Implicits.global
-import scala.concurrent.duration._
-import java.util.concurrent.TimeUnit._
-import scala.language.postfixOps
 import scala.sys.process._
+import scala.migrations.core.tools.helpers.Helpers._
 import scala.migrations.MigrationDatabase
 import scala.migrations.core.tools.{GitUtil => Git}
 import scala.migrations.tools.git.Installer
@@ -60,33 +53,11 @@ class MyMigrationDatabase(dbLoc: String, objLoc: String)
     }
   }
 
-  private def runCommandUntilNoOutput(command: Seq[String]) {
-    @tailrec def noOutput(flag: Boolean, input: InputStream, text: String): Unit = {
-      var byte: Array[Byte] = new Array(1)
-      val f = Future {
-        if (input.read(byte) != -1) ()
-      }
-      val fl = text.endsWith("Waiting for source changes... (press enter to interrupt)\n")
-      val duration = if (fl) 2 seconds else 30 seconds
-      val r = Try { Await.result(f, duration) }
-      val s = new String(byte)
-      print(s)
-      r match {
-        case Success(_) => noOutput(fl, input, text + s)
-        case Failure(ex) => ex match {
-          case te: TimeoutException =>
-            if (fl) input.close() else throw te
-        }
-      }
-    }
-
-    command run new ProcessIO(_.close(), noOutput(false, _, ""), _.close())
-  }
 
   def rebuild() {
     Seq("sbt", "mg reset").!
     Seq("sbt", "mg init").!
-    runCommandUntilNoOutput(Seq("sbt", "~mg migrate"))
+    Seq("sbt", "~mg migrate").!->
   }
 
   def rebuild(branch: String, commitId: String) {
